Move home nav components into their own folder
